refactor: separate large auth route file into smaller files
